refactor(test): Extract shared test helpers to core_test.go
- Move testPaths and calculateTestChecksum to core_test.go
- Add NewTestPaths constructor for cleaner test setup
- Remove duplicate imports and definitions
- Follow Go conventions for package-wide test helpers

This makes test helpers reusable across the core package tests.
