+++ pkg/core/pack_status_test.go
@@ -1,8 +1,6 @@
 package core
 
 import (
-	"crypto/sha256"
-	"encoding/hex"
 	"path/filepath"
 	"testing"
 	"time"
@@ -352,9 +350,7 @@ func TestGetPackStatus(t *testing.T) {
 			dataDir := "data/dodot"
 
 			// Setup paths
-			paths := &testPaths{
-				dataDir: dataDir,
-			}
+			paths := NewTestPaths(dataDir)
 
 			// Run test setup
 			if tt.setupFS != nil {
@@ -377,35 +373,4 @@ func TestGetPackStatus(t *testing.T) {
 			}
 		})
 	}
-}
-
-// testPaths implements types.Pather for testing
-type testPaths struct {
-	dataDir string
-}
-
-func (p *testPaths) DotfilesRoot() string {
-	return "dotfiles"
-}
-
-func (p *testPaths) DataDir() string {
-	return p.dataDir
-}
-
-func (p *testPaths) ConfigDir() string {
-	return filepath.Join(p.dataDir, "config")
-}
-
-func (p *testPaths) CacheDir() string {
-	return filepath.Join(p.dataDir, "cache")
-}
-
-func (p *testPaths) StateDir() string {
-	return filepath.Join(p.dataDir, "state")
-}
-
-// calculateTestChecksum calculates SHA256 checksum for test data
-func calculateTestChecksum(data []byte) string {
-	hash := sha256.Sum256(data)
-	return hex.EncodeToString(hash[:])
 }
